pyproject.toml
[project]
name = "poker-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=2.4.4",
    "requests>=2.33.1",
    "rlcard[torch]>=1.2.0",
    "torch>=2.11.0",
    "treys>=0.1.8",
    "wandb>=0.26.0",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/poker_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: slow tests requiring full training (e.g. Nash convergence)",
    "live: tests that hit live external services (e.g. slumbot.com); skipped by default — run with `uv run pytest -m live`",
]
addopts = "-m 'not live'"

[tool.ruff]
src = ["src"]
line-length = 100

[tool.mypy]
mypy_path = "src"
python_version = "3.11"
strict = true
